v5.0.0 crypto, schema refactor

+++ src/core/components/lists/SchemaList.tsx
@@ -20,41 +20,40 @@ import DeleteIcon from "@mui/icons-material/Delete";
 import EditIcon from "@mui/icons-material/Edit";
 import AppSnackbar from "../AppSnackbar";
 import { EditModalRef, EditModal } from "../forms/EditModal";
-import { ListObjectEntity } from "../forms/ObjectEntityType";
-import { FormSchema } from "../forms/SchemaForm";
+import { ListFormSchema, SchemaFormStates } from "../forms/SchemaForm";
 import ListItemLink from "../ListItemLink";
-import listReducer from "./ListReducer";
+import SchemaListReducer from "./ListReducer";
 import { AuthContext } from "../../Auth";
 
 interface SchemaListProps<T> {
   rows: T[];
   title: string;
-  getEntitySchema(obj?: T): FormSchema<T>;
-  setNewRow(obj?: T): T;
-  deleteEntity(obj: T): Promise<void>;
+  schema: ListFormSchema<T>;
+  onRowEdit?(state: SchemaFormStates, obj?: T): T;
   onChange?: (
-    schema: FormSchema<T>,
-    obj: { [key: string]: any },
-    changeObj: { [key: string]: any }
-  ) => Promise<FormSchema<T> | undefined>;
+    schema: ListFormSchema<T>,
+    obj: T,
+    changeObj: Partial<T>
+  ) => Promise<ListFormSchema<T> | undefined>;
 }
 
-function SchemaList<T extends ListObjectEntity>({
+function SchemaList<T>({
   title,
-  getEntitySchema,
-  deleteEntity,
-  setNewRow,
+  schema,
+  onRowEdit,
   rows,
   onChange,
 }: SchemaListProps<T>) {
   const authContext = useContext(AuthContext);
   const modalRef = useRef<EditModalRef>(null);
-  const [state, dispatch] = useReducer(listReducer<T>(), []);
+  const [state, dispatch] = useReducer(SchemaListReducer<T>(), []);
   const [appMessage, setAppMessage] = React.useState("");
 
-  const [schema, setEditSchema] = useState<FormSchema<T>>(() =>
-    getEntitySchema()
-  );
+  const [editSchema, setEditSchema] = useState<ListFormSchema<T>>(() => schema);
+  const [editRow, setEditRow] = useState<T>({} as T);
+  const [editState, setEditState] = useState<SchemaFormStates | undefined>();
+
+  useEffect(() => setEditSchema(schema), [schema]);
 
   useEffect(() => {
     dispatch({
@@ -64,42 +63,47 @@ function SchemaList<T extends ListObjectEntity>({
   }, [rows]);
 
   function handleDelete(row: T) {
-    deleteEntity(row)
-      .then(() => {
-        setAppMessage("Entity deleted.");
-        dispatch({ type: "DELETE", row: row });
-      })
-      .catch((err) => {
-        console.error(err);
-        setAppMessage("Delete failed, unexpected error.");
-      });
+    if (editSchema.delete) {
+      editSchema
+        .delete(row)
+        .then(() => {
+          setAppMessage("Entity deleted.");
+          dispatch({ type: "DELETE", row: row });
+        })
+        .catch((err) => {
+          console.error(err);
+          setAppMessage("Delete failed, unexpected error.");
+        });
+    }
   }
 
-  function handleEdit(row?: T) {
+  async function startEdit(state: SchemaFormStates, row?: T) {
+    row = onRowEdit ? onRowEdit(state, row) : row;
     if (row && onChange) {
-      const schema = getEntitySchema(row);
-      const asyncNewSchema = onChange(schema, schema.object, schema.object);
-      asyncNewSchema.then((newSchema) =>
-        newSchema ? setEditSchema(newSchema) : setEditSchema(schema)
-      );
-    } else {
-      setEditSchema(getEntitySchema(row));
+      const newSchema = await onChange(schema, row, row);
+      if (newSchema) {
+        setEditSchema(newSchema);
+      }
     }
+    setEditRow(row ?? ({} as T));
+    setEditState(state);
     if (modalRef && modalRef.current) {
       modalRef.current.handleOpen();
     }
   }
 
   function handleOnEditSaveSuccess(row: T) {
-    row = setNewRow(row);
-    if (schema.type === "ADD") {
+    if (editState === "ADD") {
       setAppMessage("Entity added.");
       dispatch({ type: "ADD", row: row });
-    } else if (schema.type === "EDIT") {
+    } else if (editState === "EDIT") {
       setAppMessage("Entity saved.");
       dispatch({ type: "EDIT", row: row });
     }
-    getEntitySchema();
+    setEditSchema(schema);
+    if (schema.repeatAdd && editState === "ADD") {
+      startEdit("ADD").then();
+    }
   }
 
   /**
@@ -108,14 +112,15 @@ function SchemaList<T extends ListObjectEntity>({
    * @param obj
    * @param changeObj
    */
-  function handleOnFormChange(
-    obj: { [key: string]: any },
-    changeObj: { [key: string]: any }
-  ): void {
+  function handleOnFormChange(obj: T, changeObj: Partial<T>): void {
     if (onChange) {
-      const asyncNewSchema = onChange(schema as FormSchema<T>, obj, changeObj);
+      const asyncNewSchema = onChange(
+        schema as ListFormSchema<T>,
+        obj,
+        changeObj
+      );
       asyncNewSchema.then((newSchema) =>
-        newSchema !== undefined ? setEditSchema(newSchema) : undefined
+        newSchema ? setEditSchema(newSchema) : newSchema
       );
     }
   }
@@ -126,32 +131,41 @@ function SchemaList<T extends ListObjectEntity>({
         {title}
       </Typography>
       <List component="div" dense>
-        {state.map((o) => (
-          <ListItem key={o.id}>
-            <ListItemIcon>
-              <ArrowRightIcon color="inherit" />
-            </ListItemIcon>
-            <ListItemText>
-              {o.path && <ListItemLink path={o.path} name={o.name!} />}
-              {!o.path && o.name}
-            </ListItemText>
-            {authContext.authenticated && (
-              <Fragment>
-                <IconButton onClick={() => handleEdit(o)} size="large">
-                  <EditIcon color="inherit" />
-                </IconButton>
-                <IconButton onClick={() => handleDelete(o)} size="large">
-                  <DeleteIcon color="inherit" />
-                </IconButton>
-              </Fragment>
-            )}
-          </ListItem>
-        ))}
+        {state
+          .map((o) => ({ o: o, display: schema.display(o) }))
+          .map(({ o, display }) => (
+            <ListItem key={display.id}>
+              <ListItemIcon>
+                <ArrowRightIcon color="inherit" />
+              </ListItemIcon>
+              <ListItemText>
+                {display.path && (
+                  <ListItemLink path={display.path} name={display.name!} />
+                )}
+                {!display.path && display.name}
+              </ListItemText>
+              {authContext.authenticated && (
+                <Fragment>
+                  <IconButton
+                    onClick={() => startEdit("EDIT", o).then()}
+                    size="large"
+                  >
+                    <EditIcon color="inherit" />
+                  </IconButton>
+                  <IconButton onClick={() => handleDelete(o)} size="large">
+                    <DeleteIcon color="inherit" />
+                  </IconButton>
+                </Fragment>
+              )}
+            </ListItem>
+          ))}
       </List>
       <AppSnackbar message={appMessage} onClose={() => setAppMessage("")} />
       <EditModal
         ref={modalRef}
-        schema={schema}
+        schema={editSchema}
+        obj={editRow}
+        editState={editState!}
         onSaveSuccess={handleOnEditSaveSuccess}
         onChange={handleOnFormChange}
       />
@@ -160,7 +174,7 @@ function SchemaList<T extends ListObjectEntity>({
         <Button
           color="secondary"
           variant="contained"
-          onClick={() => handleEdit()}
+          onClick={() => startEdit("ADD").then()}
         >
           Add
         </Button>
